Delete single test and test exclusion strategies for Nopol.
