+++ src/repairnator-jenkins-plugin/src/main/resources/io/jenkins/plugins/main/RepairnatorPostBuild/config.jelly
@@ -21,9 +21,6 @@
         <f:entry name="useAstorJKali" title="${%AstorJKali}" field="useAstorJKali">
             <f:checkbox />
         </f:entry>
-        <f:entry name="useNopolTestExclusionStrategy" title="${%NopolTestExclusionStrategy}" field="useNopolTestExclusionStrategy">
-            <f:checkbox />
-        </f:entry>
         <f:entry name="useSorald" title="${%Sorald}" field="useSorald">
             <f:checkbox />
         </f:entry>
