+++ src/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/inspectors/TestProjectInspector.java
@@ -61,7 +61,7 @@ public class TestProjectInspector {
         config.setZ3solverPath(Utils4Tests.getZ3SolverPath());
         config.setPush(true);
         config.setPushRemoteRepo("");
-        config.setRepairTools(new HashSet<>(Arrays.asList("NPEFix", "NopolSingleTest")));
+        config.setRepairTools(new HashSet<>(Arrays.asList("NPEFix", "Nopol")));
         config.setGithubUserEmail("noreply@github.com");
         config.setGithubUserName("repairnator");
         Utils.setLoggersLevel(Level.ERROR);
