+++ src/repairnator-pipeline/src/test/java/fr/inria/spirals/repairnator/process/inspectors/metrics4bears/TestMetrics4BearsJsonFile.java
@@ -185,7 +185,7 @@ public class TestMetrics4BearsJsonFile {
 
         RepairnatorConfig config = RepairnatorConfig.getInstance();
         config.setLauncherMode(LauncherMode.REPAIR);
-        config.setRepairTools(new HashSet<>(Arrays.asList("NopolSingleTest")));
+        config.setRepairTools(new HashSet<>(Arrays.asList("Nopol")));
 
         ProjectInspector inspector = new ProjectInspector(buildToBeInspected, tmpDir.getAbsolutePath(), null, null);
         inspector.run();
