+++ src/scripts/config/repairnator.cfg
@@ -20,7 +20,7 @@ SMTP_PASSWORD= #Password for authorized server, if not left empty also enables a
 NOTIFY_TO= # email adresses separated by comma (used for notifier, you must have specify an smtp server)
 NOTIFY_ENDPROCESS=0 # If set to 1, the end of dockerpool and scanner will send a notification using smtp and notify_to information
 RUN_ID_SUFFIX= # A suffix to add to the run id (useful to retrieve a specific run in serialized data).
-REPAIR_TOOLS=NopolAllTests,AstorJMut,NPEFix # The available repair tools to use. Possible values (comma separated): NPEFix, AssertFixer, AstorJMut, AstorJKali, AstorJGenProg, NopolAllTests, NopolSingleTest, NopolTestExclusionStrategy
+REPAIR_TOOLS=Nopol,AstorJMut,NPEFix # The available repair tools to use. Possible values (comma separated): NPEFix, AssertFixer, AstorJMut, AstorJKali, AstorJGenProg, Nopol
 
 CREATE_PR=0 # set to 1 to create pull requests automatically when a patch is found
